--- lib-multisrc/mayotune/build.gradle.kts
@@ -1,5 +1,0 @@
-plugins {
-    id("lib-multisrc")
-}
-
-baseVersionCode = 1
