+++ src/en/mayotune/res/mipmap-mdpi/ic_launcher.png
